requirements.txt
anthropic>=0.92.0
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.6
sse-starlette>=2.0
gradio>=4.44
numpy>=1.24
pillow>=10.0
opencv-python-headless>=4.8
matplotlib>=3.7
reportlab>=4.0
pydantic>=2.0
python-dotenv>=1.0
pandas>=2.0
pytest>=7.4

# Uncomment to enable real-model inference / training:
# tensorflow>=2.13
# scikit-learn>=1.3
